requirements.txt
diffusers>=0.32.0,<1.0
transformers>=4.40.0,<5.0
accelerate>=0.30.0,<2.0
safetensors>=0.4.0,<1.0

Pillow>=10.0.0,<12.0
numpy>=1.24.0,<3.0
scipy>=1.11.0,<2.0
opencv-python-headless>=4.8.0,<5.0

matplotlib>=3.7.0,<4.0

peft>=0.10.0,<1.0

# huggingface-hub>=0.23.0,<1.0

mypy>=1.10.0
types-Pillow